Move highlighter to a separate file
